refactor: extract Standings interface
